# Project management application/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0", 
    "title": "Manager app backend",
    "description": "Manager Application API",
    "license": "ISC"
  },
...
  "securityDefinitions":{
    "Bearer":{
      "type": "apiKey",
      "name": "Authorization",
      "description": "Enter your token in format *Bearer token*",
      "in": "header"
    }
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths":{
...
    "/boards/{boardId}/columns": {
      "get": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Get Columns in Board",
        "description": "Get Columns in board",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "All columns in board",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Column"
              } 
            }
          }
        }
      },
      "post": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Create Column",
        "description": "Create Column in board",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "new Column params",
            "in": "body",
            "description": "New params",
            "schema": {
              "required": [
                "title",
                "order"
              ],
              "properties": {
                "title": {
                  "type": "string"
                },
                "order": {
                  "type": "number"
                }      
              }
            }
            
          }
        ],
        "responses": {
          "200": {
            "description": "Created column",
            "schema": {
              "$ref": "#/definitions/Column"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"              
            }
          }
        }
      }
    },    
    "/boards/{boardId}/columns/{columnId}": {
      "get": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Get Column by Id",
        "description": "Find Column",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Founded column",
            "schema": {
              "$ref": "#/definitions/Column"
            }
          },
          "404": {
            "description": "Column was not founded!",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Update Column by Id",
        "description": "Update Column",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          },          
          {
            "name": "new Column params",
            "in": "body",
            "description": "New params",
            "schema": {
              "required": [
                "title",
                "order"
              ],
              "properties": {
                "title": {
                  "type": "string"
                },
                "order": {
                  "type": "number"
                }       
              }
            }
            
          }
        ],
        "responses": {
          "200": {
            "description": "Updated column",
            "schema": {
              "$ref": "#/definitions/Column"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"              
            }
          }
        }
      },
      "delete": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Columns"],
        "summary": "Delete Column by Id",
        "description": "Delete Column",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted column",
            "schema": {
              "$ref": "#/definitions/Column"
            }
          }
        }
      }
    },
...
    "/boards/{boardId}/columns/{columnId}/tasks": {
      "get": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Tasks"],
        "summary": "Get Tasks in Column",
        "description": "Get Tasks in columns",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "All tasks in column",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Task"
              } 
            }
          }
        }
      },
      "post": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Tasks"],
        "summary": "Create Task",
        "description": "Create Task in column",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "new Task params",
            "in": "body",
            "description": "New params",
            "schema": {
              "required": [
                "title",
                "order",
                "description",
                "userId",
                "users"
              ],
              "properties": {
                "title": {
                  "type": "string"
                },
                "order": {
                  "type": "number"
                },
                "description": {
                  "type": "string"
                },
                "userId": {
                  "type": "number"
                },
                "users": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  } 
                }       
              }
            }
            
          }
        ],
        "responses": {
          "200": {
            "description": "Created task",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"              
            }
          }
        }
      }
    },    
    "/boards/{boardId}/columns/{columnId}/tasks/{taskId}": {
      "get": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Tasks"],
        "summary": "Get Task by Id",
        "description": "Find Task",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Founded task",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          "404": {
            "description": "Task was not founded!",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Tasks"],
        "summary": "Update Task by Id",
        "description": "Update Task",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "new Task params",
            "in": "body",
            "description": "New params",
            "schema": {
              "required": [
                "title",
                "order",
                "description",
                "columnId",
                "userId",
                "users"
              ],
              "properties": {
                "title": {
                  "type": "string"
                },
                "order": {
                  "type": "number"
                },
                "description": {
                  "type": "string"
                },
                "columnId": {
                  "type": "string"
                },
                "userId": {
                  "type": "number"
                },
                "users": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  } 
                }       
              }
            }
            
          }
        ],
        "responses": {
          "200": {
            "description": "Updated task",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"              
            }
          }
        }
      },
      "delete": {
        "security": [
          {
             "Bearer": []
          }
        ],
        "tags": ["Tasks"],
        "summary": "Delete Task by Id",
        "description": "Delete Task",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "description": "Board ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "columnId",
            "in": "path",
            "description": "Column ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted task",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Column": {
      "required": [
        "_id",
        "title",
        "order",
        "boardId"
      ],
      "properties": {
        "_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "order": {
          "type": "number"
        },
        "boardId": {
          "type": "string"
        }      
      },
      "example": {
        "_id": "Column id",
        "title": "Column title",
        "order": 1,
        "boardId": "Id of boards"
      }
    },
    "Task": {
      "required": [
        "_id",
        "title",
        "order",
        "boardId",
        "columnId",
        "description",
        "userId",
        "users"
      ],
      "properties": {
        "_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "order": {
          "type": "number"
        },
        "boardId": {
          "type": "string"
        },
        "columnId": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "userId": {
          "type": "number"
        },
        "users": {
          "type": "array",
          "items": {
            "type": "string"
          } 
        }       
      },
      "example": {
        "_id": "Task id",
        "title": "Task title",
        "order": 0,
        "boardId": "id of board",
        "columnId": "id of column",
        "description": "Task decription",
        "userId": "userId of task owner",
        "users": [
          "userId of responsible user #1",
          "userId of responsible user #2"
        ]
      }
    },
...
    "ErrorResponse": {
      "required": [
        "statusCode",
        "message"
      ],
      "properties": {
        "statusCode": {
          "type": "number"
        },
        "message": {
          "type": "string"
        }     
      },
      "example": {
        "statusCode": "4xx",
        "message": "reason"
      }
    }
  }
}